Fix: Review  -Separate openBrowser into own file
